+++ ibis/backends/clickhouse/operations.py
@@ -6,11 +6,11 @@ import ibis.expr.operations as ops
 import ibis.expr.types as ir
 import ibis.sql.transforms as transforms
 import ibis.util as util
-from ibis.clickhouse.identifiers import quote_identifier
+from ibis.backends.clickhouse.identifiers import quote_identifier
 
 
 def _cast(translator, expr):
-    from ibis.clickhouse.client import ClickhouseDataType
+    from ibis.backends.clickhouse.client import ClickhouseDataType
 
     op = expr.op()
     arg, target = op.args
